+++ notifications_utils/template.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from functools import lru_cache
 from html import unescape
 from os import path
+from typing import Optional
 
 from jinja2 import Environment, FileSystemLoader
 from markupsafe import Markup
@@ -41,9 +42,11 @@ from notifications_utils.markdown import (
     notify_email_markdown,
     notify_email_preheader_markdown,
     notify_letter_preview_markdown,
+    notify_letter_qrcode_validator,
     notify_plain_text_email_markdown,
 )
 from notifications_utils.postal_address import PostalAddress, address_lines_1_to_7_keys
+from notifications_utils.qr_code import QrCodeTooLong
 from notifications_utils.sanitise_text import SanitiseSMS
 from notifications_utils.take import Take
 from notifications_utils.template_change import TemplateChange
@@ -704,6 +707,15 @@ class BaseLetterTemplate(SubjectMixin, Template):
     def postal_address(self):
         return PostalAddress.from_personalisation(InsensitiveDict(self.values))
 
+    def has_qr_code_with_too_much_data(self) -> Optional[QrCodeTooLong]:
+        content = self._personalised_content if self.values else self.content
+        try:
+            Take(content).then(notify_letter_qrcode_validator)
+        except QrCodeTooLong as e:
+            return e
+
+        return None
+
     @property
     def _address_block(self):
 
@@ -740,17 +752,19 @@ class BaseLetterTemplate(SubjectMixin, Template):
         return self.date.strftime("%-d %B %Y")
 
     @property
+    def _personalised_content(self) -> Field:
+        return Field(
+            self.content,
+            self.values,
+            html="escape",
+            markdown_lists=True,
+            redact_missing_personalisation=self.redact_missing_personalisation,
+        )
+
+    @property
     def _message(self):
         return (
-            Take(
-                Field(
-                    self.content,
-                    self.values,
-                    html="escape",
-                    markdown_lists=True,
-                    redact_missing_personalisation=self.redact_missing_personalisation,
-                )
-            )
+            Take(self._personalised_content)
             .then(add_trailing_newline)
             .then(notify_letter_preview_markdown)
             .then(do_nice_typography)