+++ CHANGELOG.md
@@ -7,6 +7,23 @@ More minor changes may optionally be recorded here too.
 
 * InvalidPhoneError messages have been updated. The new error messages are more user friendly.
 
+## 69.0.0
+
+* Remove old syntax for QR codes in letters (only `QR: http://example.com` will work now)
+
+## 68.0.1
+
+* Fix a bug with some HTML getting injected into QR codes
+
+## 68.0.0
+
+* Update return value of `BaseLetterTemplate.has_qr_code_with_too_much_data` from `bool` to `Optional[QrCodeTooLong]`.
+
+## 67.0.0
+
+* Add `has_qr_code_with_too_much_data` property to letter templates.
+* Update RecipientCSV to detect and throw errors when rows generate QR codes with too much data in them. Anything using RecipientCSV to process letters will need to check for and report on the row-level property `qr_code_too_long`.
+
 ## 66.1.0
 
 * Add a simpler syntax for QR codes in letters (QR: http://example.com)